Check that the error is raised
It seemed unnecessairly complex to trap the error and report a failure
to the 'user' given the simplistic nature of the dummy app.  Just seeing
that the error is raised feels sufficient.

I decided to remove the confirmation_code error given that we are still
discussing this and, right now, GLiMR does not implement it in any
meaningful way.  I can put it back in once we have functionality to test
against.