Enable and improve Code Climate metrics
I added the repo to Code Climate in an effort to get C0 coverage
working. SimpleCov refuses to pick up the contents of `lib` regardless
of the gem-coverage-problem workarounds I have tried.  Adding Code
Climate did not help resolve the problem as I remembered, too late, that
Code Climate only reports on the results of whatever coverage tool is
being used locally.  I decided to leave Code Coverage in as it is
otherwise unproblematic and several people in the department like and
use it. I completely removed SimpleCov for now.

The details on the simplecov issue along with the proposed, non-working,
workaround can be found here:

https://github.com/colszowka/simplecov/issues/221

Please see the comment on mutation testing for furhter details on
coverage.