Stop mutation tests from breaking for now
I need to properly stub out the connection_timeout to keep it from being
susceptible to mutation.  However, the higher priority at the moment is
getting the CI working again so that master branch protection can be
turned on on github.  This resolves the problem by going back to the
default timeout.  I'll redo this again in a feature branch at a future
date.