Prevent integration tests from breaking (#11)
* Prevent integration tests from breaking

The `DEFAULT_URL` constant wasn't picking up the value of
`GLIMR_API_URL` when the gem was being run in integration tests in
`tax-tribunals-fees`.  This was because the `dotenv` gem had to be
loaded and required prior to the `glimr-api-client` gem. In order to
avoid doing this, I have changed the constant to a method that fetches
the value when it is called.

* fixup! Prevent integration tests from breaking

* I missed the updated to the Gemfile.lock

Bumping the version as this went out in the published gem as well.