Json confirmation code fix (#12)
* Use a constant for jurisdictionId

...because magic numbers are icky

* Enable confirmation_code parameter

The confirmation_code was not passed to glimr in the first version of
the code for reasons that are not clear at this point. This commit
changes the behaviour in order to send the confirmation code to glimr.

* Change request bodies to JSON

Previously, glimr accepted query strings posted with a mime type of
`application/json`. This broke with a recent release and this commit
fixes that.

* ignore development artefacts

* Do not require active_support

If you do, then running the gem in a context where
rails is not available will raise a LoadError

* update rails gems

This was a by-product of reviewing the latest gem
changes

* require two parameters when finding a case

* remove todo comment - it's done