Stop excon web stub from hooking silently
When porting the validation code to PBA I discovered this spec was
passing because the Excon web stub was hooking in the validating
parameters context.  The before block that *seems* to stub the client
with a double does not work as it is stubbed on the wrong instance.

The spec was still passing because of the shared examples, which
inclded an Excon web stub that *was* hooking and happened to work.

I've changed this to use the (frankly smellier) `allow_any_instance_of`
stub, which makes all the validation specs use the stubs defined at the
beginning of this spec.  I also removed the shared examples altoghether
as they are no longer needed.