Add pba method (#15)
* Enable posts to PayByAccount

This is a new endpoint on GLiMR that allows users to pay by account.

* Stop excon web stub from hooking silently

When porting the validation code to PBA I discovered this spec was
passing because the Excon web stub was hooking in the validating
parameters context.  The before block that *seems* to stub the client
with a double does not work as it is stubbed on the wrong instance.

The spec was still passing because of the shared examples, which
inclded an Excon web stub that *was* hooking and happened to work.

I've changed this to use the (frankly smellier) `allow_any_instance_of`
stub, which makes all the validation specs use the stubs defined at the
beginning of this spec.  I also removed the shared examples altoghether
as they are no longer needed.

* Validate client data

Ensure the caller sends the correct information. This is becoming our
standard pattern for this type of spec.

* Document the RequestNewCase call

This is done in the simplest possible way.

* Document PBA method call in README

* List methods alphabetically

I had added them in no particular order, despite the existing ones being
listed alphabetically.

* Make consistent with PBA

No need for a method to expose the params, an attribute reader works
fine.

* Clairify the response_body values

This *should* obviate the need for a feature spec.

* Reraise the GLiMR error from the spec

This covers all of them for PBA.

* Refactor error handling

Break it out into individual classes

* Bring Update in-line with other classes

See it and Case for specific TODOs.

* Ease deployment issues with sensible timeout

The default Excon timeout is 60 seconds.  This is causing problems in
the deployment pipeline when GLiMR isn't available.  This reduces the
default to five seconds with an override set via environment variable.

* Simplify error handling reduce spec fragility

API specific errors are now handled in their respective classes with a
fallback to a generic `Unavailable` error if nothing else hooks. The
dummy rails app is gone and the specs have been standardised.

* Stop circle breaking

It no longer needs a database.

* Stop circle attempting db install

* Circle is still trying to run rake db:create

* Add attribute missing from docs

* typo: Nubmer -> Number