Longer timeout for register (#22)
* Bring in line with other TT services

This was locked on 2.3.1, but should have been 2.3.3.  Updated the gem
dependencies at the same time.

* Prep for changeover to travis

Also part of the work to bring it in line with all the other TaxTribs
applications.

* Fix the timeout issues on prod/staging

On reflection, setting the `read_timeout` probably wasn't the correct
thing to do...Changed for `write_timeout` and `connection_timeout` and
ensured that `RegisterNewCase` has a timeout that reflects the runtimes
we have seen on the running service (see comments in the code for
details).

* Reduce maintenance overhead

Owing to the ministerial delays regarding fees and the just-called
general election, this will not be required for at least several months.
Removing it for now as it does nothing but add to the overhead of
maintaining the gem.

* Use a REST client with thread-safe timeouts

Excon timeouts were not working as expected/advertised.  Typhoeus is a
set of wrappers around libcurl and manual testing has confirmed that its
timeout setting work properly in a threaded environment.

This commit also brings the gem versions up-to-date.

* Swap build badge from circle to travis

* Minor version bump

On reflection, sufficient amounts have chagned that a larger bump seemed
sensible.