Restore ability to find cases
`::Case` was only used for fee payments and was removed at the same time
I changed the REST client over to Typheous. After discussions with DS,
I've added it back in ahead of the possibility of the fees being
re-activated in the main app.

As the app will need to be adapted anyway, I have taken this opportunity
to rename `Case` to the clearer `FindCase`, which is something that had
been on the tech debt todo list for a long time.