Restore payment calls (#23)
* Restore ability to find cases

`::Case` was only used for fee payments and was removed at the same time
I changed the REST client over to Typheous. After discussions with DS,
I've added it back in ahead of the possibility of the fees being
re-activated in the main app.

As the app will need to be adapted anyway, I have taken this opportunity
to rename `Case` to the clearer `FindCase`, which is something that had
been on the tech debt todo list for a long time.

* Do not error on empty response bodies

This was relevant for `RegisterNewCase`.  However, some of the payment
calls (HWF, for example) may return empty reponse bodies.

* Remove mutation point

`#re_raise_error` cannot receive a body without a :message at this
level.

* Restore ability to enter Help With Fees reference

`::HwfRequested` was only used for fee payments and was removed at the
same time I changed the REST client over to Typheous. After discussions
with DS, I've added it back in ahead of the possibility of the fees
being re-activated in the main app.

* Restore ability to Pay By Account

`::PayByAccount` was only used for fee payments and was removed at the
same time I changed the REST client over to Typheous. After discussions
with DS, I've added it back in ahead of the possibility of the fees
being re-activated in the main app.

* Restore ability to update GLiMR with payment

`::Update` was only used for fee payments and was removed at
the same time I changed the REST client over to Typheous. After
discussions with DS, I've added it back in ahead of the possibility of
the fees being re-activated in the main app.

* Explain reasons behind spec layout

* Bump version